Refactor complete, but tests failing for AlphaBetaSearch

+++ src/model/ai/evaluation/BoardEvaluatorSummator.java
@@ -13,11 +13,11 @@ public class BoardEvaluatorSummator implements BoardEvaluatorAggregator {
     public double evaluateBoard(Strategy strategy, Board theBoard) {
         double value = 0.0;
         final PieceColor color = strategy.getColor();
-        final HashMap<BoardEvaluatorInterface, Float> weightMap = strategy.getEvaluatorWeightMap();
+        final HashMap<BoardEvaluatorInterface, Double> weightMap = strategy.getEvaluatorWeightMap();
 
-        for (final Map.Entry<BoardEvaluatorInterface, Float> entry : weightMap.entrySet()) {
+        for (final Map.Entry<BoardEvaluatorInterface, Double> entry : weightMap.entrySet()) {
             final BoardEvaluatorInterface evaluator = entry.getKey();
-            final float weight = entry.getValue().floatValue();
+            final double weight = entry.getValue().doubleValue();
             value += evaluator.evaluateBoard(theBoard, color) * weight;
         }
